+++ Sources/Rendering/WebGPU/FullScreenQuad/index.js
@@ -1,6 +1,6 @@
 import macro from 'vtk.js/Sources/macros';
 import vtkWebGPUShaderCache from 'vtk.js/Sources/Rendering/WebGPU/ShaderCache';
-import vtkWebGPUMapperHelper from 'vtk.js/Sources/Rendering/WebGPU/MapperHelper';
+import vtkWebGPUSimpleMapper from 'vtk.js/Sources/Rendering/WebGPU/SimpleMapper';
 
 // ----------------------------------------------------------------------------
 // vtkWebGPUFullScreenQuad methods
@@ -25,12 +25,10 @@ function vtkWebGPUFullScreenQuad(publicAPI, model) {
     publicAPI.replaceShaderPosition
   );
 
-  const superclassBuild = publicAPI.build;
-  publicAPI.build = (renderEncoder, device) => {
-    const buff = device.getBufferManager().getFullScreenQuadBuffer();
+  publicAPI.updateBuffers = () => {
+    const buff = model.device.getBufferManager().getFullScreenQuadBuffer();
     model.vertexInput.addBuffer(buff, ['vertexBC']);
     model.numberOfVertices = 6;
-    superclassBuild(renderEncoder, device);
   };
 }
 
@@ -46,7 +44,7 @@ export function extend(publicAPI, model, initialValues = {}) {
   Object.assign(model, DEFAULT_VALUES, initialValues);
 
   // Inheritance
-  vtkWebGPUMapperHelper.extend(publicAPI, model, initialValues);
+  vtkWebGPUSimpleMapper.extend(publicAPI, model, initialValues);
 
   // Object methods
   vtkWebGPUFullScreenQuad(publicAPI, model);
