+++ Sources/Rendering/WebGPU/SphereMapper/index.js
@@ -1,12 +1,12 @@
 import * as macro from 'vtk.js/Sources/macros';
 import * as vtkMath from 'vtk.js/Sources/Common/Core/Math';
-import vtkWebGPUPolyDataMapper from 'vtk.js/Sources/Rendering/WebGPU/PolyDataMapper';
+import vtkWebGPUCellArrayMapper from 'vtk.js/Sources/Rendering/WebGPU/CellArrayMapper';
 import vtkWebGPUBufferManager from 'vtk.js/Sources/Rendering/WebGPU/BufferManager';
 import vtkWebGPUShaderCache from 'vtk.js/Sources/Rendering/WebGPU/ShaderCache';
 
 import { registerOverride } from 'vtk.js/Sources/Rendering/WebGPU/ViewNodeFactory';
 
-const { BufferUsage, PrimitiveTypes } = vtkWebGPUBufferManager;
+const { BufferUsage } = vtkWebGPUBufferManager;
 const { vtkErrorMacro } = macro;
 
 const vtkWebGPUSphereMapperVS = `
@@ -65,6 +65,21 @@ fn main(
 function vtkWebGPUSphereMapper(publicAPI, model) {
   // Set our className
   model.classHierarchy.push('vtkWebGPUSphereMapper');
+
+  const cellMapperBuildPass = publicAPI.buildPass;
+  publicAPI.buildPass = (prepass) => {
+    if (prepass) {
+      if (!model.renderable.getStatic()) {
+        model.renderable.update();
+      }
+
+      const poly = model.renderable.getInputData();
+
+      publicAPI.setCellArray(poly.getVerts());
+      publicAPI.setCurrentInput(poly);
+    }
+    cellMapperBuildPass(prepass);
+  };
 
   publicAPI.replaceShaderNormal = (hash, pipeline, vertexInput) => {
     const vDesc = pipeline.getShaderDescription('vertex');
@@ -140,37 +155,28 @@ function vtkWebGPUSphereMapper(publicAPI, model) {
   // compute a unique hash for a pipeline, this needs to be unique enough to
   // capture any pipeline code changes (which includes shader changes)
   // or vertex input changes/ bind groups/ etc
-  publicAPI.computePipelineHash = (vertexInput) => {
-    let pipelineHash = 'spm';
-    if (vertexInput.hasAttribute(`colorVI`)) {
-      pipelineHash += `c`;
+  publicAPI.computePipelineHash = () => {
+    model.pipelineHash = 'spm';
+    if (model.vertexInput.hasAttribute(`colorVI`)) {
+      model.pipelineHash += `c`;
     }
-    pipelineHash += model.renderEncoder.getPipelineHash();
-
-    return pipelineHash;
+    model.pipelineHash += model.renderEncoder.getPipelineHash();
   };
 
-  // was originally buildIBOs() but not using IBOs right now
-  publicAPI.buildPrimitives = () => {
+  publicAPI.updateBuffers = () => {
     const poly = model.currentInput;
 
-    const device = model.WebGPURenderWindow.getDevice();
-
     model.renderable.mapScalars(poly, 1.0);
-
-    // handle triangles
-    const i = PrimitiveTypes.Triangles;
 
     const points = poly.getPoints();
     const numPoints = points.getNumberOfPoints();
     const pointArray = points.getData();
-    const primHelper = model.primitives[i];
 
     // default to one instance and computed number of verts
-    primHelper.setNumberOfInstances(1);
-    primHelper.setNumberOfVertices(3 * numPoints);
+    publicAPI.setNumberOfInstances(1);
+    publicAPI.setNumberOfVertices(3 * numPoints);
 
-    const vertexInput = model.primitives[i].getVertexInput();
+    const vertexInput = model.vertexInput;
 
     let buffRequest = {
       owner: points,
@@ -179,7 +185,7 @@ function vtkWebGPUSphereMapper(publicAPI, model) {
       usage: BufferUsage.RawVertex,
       format: 'float32x3',
     };
-    if (!device.getBufferManager().hasBuffer(buffRequest)) {
+    if (!model.device.getBufferManager().hasBuffer(buffRequest)) {
       // xyz v1 v2 v3
       const tmpVBO = new Float32Array(3 * numPoints * 3);
 
@@ -198,7 +204,7 @@ function vtkWebGPUSphereMapper(publicAPI, model) {
         tmpVBO[vboIdx++] = pointArray[pointIdx + 2];
       }
       buffRequest.nativeArray = tmpVBO;
-      const buff = device.getBufferManager().getBuffer(buffRequest);
+      const buff = model.device.getBufferManager().getBuffer(buffRequest);
       vertexInput.addBuffer(buff, ['vertexBC']);
     }
 
@@ -223,7 +229,7 @@ function vtkWebGPUSphereMapper(publicAPI, model) {
         usage: BufferUsage.RawVertex,
         format: 'float32x2',
       };
-      if (!device.getBufferManager().hasBuffer(buffRequest)) {
+      if (!model.device.getBufferManager().hasBuffer(buffRequest)) {
         const tmpVBO = new Float32Array(3 * numPoints * 2);
 
         const cos30 = Math.cos(vtkMath.radiansFromDegrees(30.0));
@@ -241,13 +247,11 @@ function vtkWebGPUSphereMapper(publicAPI, model) {
           tmpVBO[vboIdx++] = 2.0 * radius;
         }
         buffRequest.nativeArray = tmpVBO;
-        const buff = device.getBufferManager().getBuffer(buffRequest);
+        const buff = model.device.getBufferManager().getBuffer(buffRequest);
         vertexInput.addBuffer(buff, ['offsetMC']);
       }
       model._lastRadius = defaultRadius;
     }
-
-    model.renderable.mapScalars(poly, 1.0);
 
     // deal with colors but only if modified
     let haveColors = false;
@@ -261,7 +265,7 @@ function vtkWebGPUSphereMapper(publicAPI, model) {
           usage: BufferUsage.RawVertex,
           format: 'unorm8x4',
         };
-        if (!device.getBufferManager().hasBuffer(buffRequest)) {
+        if (!model.device.getBufferManager().hasBuffer(buffRequest)) {
           const colorComponents = c.getNumberOfComponents();
           if (colorComponents !== 4) {
             vtkErrorMacro('this should be 4');
@@ -279,7 +283,7 @@ function vtkWebGPUSphereMapper(publicAPI, model) {
             }
           }
           buffRequest.nativeArray = tmpVBO;
-          const buff = device.getBufferManager().getBuffer(buffRequest);
+          const buff = model.device.getBufferManager().getBuffer(buffRequest);
           vertexInput.addBuffer(buff, ['colorVI']);
         }
         haveColors = true;
@@ -289,11 +293,8 @@ function vtkWebGPUSphereMapper(publicAPI, model) {
       vertexInput.removeBufferIfPresent('colorVI');
     }
 
-    primHelper.setPipelineHash(publicAPI.computePipelineHash(vertexInput));
-    primHelper.setWebGPURenderer(model.WebGPURenderer);
-    primHelper.setTopology('triangle-list');
-    primHelper.build(model.renderEncoder, device);
-    primHelper.registerToDraw();
+    publicAPI.setTopology('triangle-list');
+    publicAPI.updateUBO();
   };
 }
 
@@ -309,16 +310,14 @@ export function extend(publicAPI, model, initialValues = {}) {
   Object.assign(model, DEFAULT_VALUES, initialValues);
 
   // Inheritance
-  vtkWebGPUPolyDataMapper.extend(publicAPI, model, initialValues);
+  vtkWebGPUCellArrayMapper.extend(publicAPI, model, initialValues);
 
-  model.primitives[PrimitiveTypes.Triangles].setVertexShaderTemplate(
-    vtkWebGPUSphereMapperVS
-  );
+  publicAPI.setVertexShaderTemplate(vtkWebGPUSphereMapperVS);
 
   // Object methods
   vtkWebGPUSphereMapper(publicAPI, model);
 
-  const sr = model.primitives[PrimitiveTypes.Triangles].getShaderReplacements();
+  const sr = model.shaderReplacements;
   sr.set('replaceShaderPosition', publicAPI.replaceShaderPosition);
   sr.set('replaceShaderNormal', publicAPI.replaceShaderNormal);
 }
